Implement bare minimum support for viewing quote posts
Displays a button that navigates to the quoted post. No preview yet. This is necessary to make any sense of posts that quote other posts on Firefish instances; otherwise, they look like reblogs, and it's very confusing.

Adds recursive account and status saves, which fixes the Firefish error 19 issue.

Fixes #124
Related to #148