Fix Siren on Catalyst
Catalyst actually can import AppKit, but we want it to use UIKit, so this swaps the conditional compilation stuff so it will check for UIKit *first* and avoid importing AppKit in that scenario. This prevents it from trying to compile against the partial AppKit available to Catalyst apps, which doesn't include text APIs.